+++ app/components/Theme.tsx
@@ -1,8 +1,8 @@
 import { observer } from "mobx-react";
 import * as React from "react";
 import { ThemeProvider } from "styled-components";
-import useMediaQuery from "~/hooks/useMediaQuery";
 import { dark, light, lightMobile, darkMobile } from "@shared/theme";
+import useMediaQuery from "~/hooks/useMediaQuery";
 import useStores from "~/hooks/useStores";
 import GlobalStyles from "~/styles/globals";
