+++ app/editor/components/WithTheme.tsx
@@ -1,5 +1,7 @@
 import * as React from "react";
 import { DefaultTheme, useTheme } from "styled-components";
+import { light } from "@shared/theme";
+import useMediaQuery from "~/hooks/useMediaQuery";
 
 type Props = {
   children: (theme: DefaultTheme) => React.ReactElement;
@@ -7,5 +9,7 @@ type Props = {
 
 export default function WithTheme({ children }: Props) {
   const theme = useTheme();
-  return children(theme);
+  const isPrinting = useMediaQuery("print");
+
+  return children(isPrinting ? light : theme);
 }
