+++ app/hooks/useMobile.ts
@@ -1,5 +1,5 @@
-import useMediaQuery from "~/hooks/useMediaQuery";
 import { base } from "@shared/theme";
+import useMediaQuery from "~/hooks/useMediaQuery";
 
 export default function useMobile(): boolean {
   return useMediaQuery(`(max-width: ${base.breakpoints.tablet - 1}px)`);
